output: merge light functions
