+++ src/libtrx/game/output/lights.c
@@ -2,13 +2,26 @@
 
 #include "game/matrix.h"
 #include "game/output.h"
+#include "game/random.h"
 #include "utils.h"
+
+#define M_LIGHT_CYCLE 32
+#define M_MAX_ROOM_LIGHT_UNIT (0x2000 / (M_LIGHT_CYCLE / 2))
 
 typedef struct {
     XYZ_32 pos;
     int32_t shade;
 } M_COMMON_LIGHT;
 
+typedef struct {
+    int32_t table[M_LIGHT_CYCLE];
+} M_ROOM_LIGHT_TABLE;
+
+static bool m_IsSunsetEnabled = false;
+static int32_t m_SunsetTimer = 0;
+static int32_t m_WibbleOffset = 0;
+static int32_t m_RoomLightShades[RLM_NUMBER_OF] = {};
+static M_ROOM_LIGHT_TABLE m_RoomLightTables[M_LIGHT_CYCLE] = {};
 static VECTOR *m_DynamicLights = nullptr;
 
 static void M_CalculateBrightestLight(
@@ -20,7 +33,6 @@ static void M_CalculateBrightestLight(
     const XYZ_32 pos, const ROOM *const room,
     M_COMMON_LIGHT *const brightest_light)
 {
-#if TR_VERSION == 2
     if (room->light_mode != RLM_NORMAL) {
         const int32_t light_shade = Output_GetRoomLightShade(room->light_mode);
         for (int32_t i = 0; i < room->num_lights; i++) {
@@ -37,8 +49,8 @@ static void M_CalculateBrightestLight(
                 falloff_1 * light->shade.value_1 / (falloff_1 + dist);
             const int32_t shade_2 =
                 falloff_2 * light->shade.value_2 / (falloff_2 + dist);
-            const int32_t shade =
-                shade_1 + (shade_2 - shade_1) * light_shade / (WIBBLE_SIZE - 1);
+            const int32_t shade = shade_1
+                + (shade_2 - shade_1) * light_shade / (M_LIGHT_CYCLE - 1);
 
             if (shade > brightest_light->shade) {
                 brightest_light->shade = shade;
@@ -47,7 +59,6 @@ static void M_CalculateBrightestLight(
         }
         return;
     }
-#endif
 
     const int32_t ambient = TR_VERSION == 1 ? (SHADE_MAX - room->ambient) : 0;
     for (int32_t i = 0; i < room->num_lights; i++) {
@@ -119,16 +130,16 @@ void Output_CalculateLight(const XYZ_32 pos, const int16_t room_num)
         global_adder = room->ambient;
         global_divider = 0;
     } else {
-#if TR_VERSION == 1
-        global_adder = SHADE_MAX - adder;
-        const int32_t divider = brightest_light.shade == adder
-            ? adder
-            : brightest_light.shade - adder;
-        global_divider = (1 << (W2V_SHIFT + 12)) / divider;
-#else
-        global_adder = room->ambient - adder;
-        global_divider = (1 << (W2V_SHIFT + 12)) / adder;
-#endif
+        if (TR_VERSION == 1) {
+            global_adder = SHADE_MAX - adder;
+            const int32_t divider = brightest_light.shade == adder
+                ? adder
+                : brightest_light.shade - adder;
+            global_divider = (1 << (W2V_SHIFT + 12)) / divider;
+        } else {
+            global_adder = room->ambient - adder;
+            global_divider = (1 << (W2V_SHIFT + 12)) / adder;
+        }
         int16_t angles[2];
         Math_GetVectorAngles(
             pos.x - brightest_light.pos.x, pos.y - brightest_light.pos.y,
@@ -158,10 +169,10 @@ void Output_CalculateStaticMeshLight(
     const XYZ_32 pos, const SHADE shade, const ROOM *const room)
 {
     int32_t adder = shade.value_1;
-    if (TR_VERSION == 2 && room->light_mode != RLM_NORMAL) {
+    if (room->light_mode != RLM_NORMAL) {
         const int32_t room_shade = Output_GetRoomLightShade(room->light_mode);
         adder +=
-            (shade.value_2 - shade.value_1) * room_shade / (WIBBLE_SIZE - 1);
+            (shade.value_2 - shade.value_1) * room_shade / (M_LIGHT_CYCLE - 1);
     }
 
     for (int32_t i = 0; i < m_DynamicLights->count; i++) {
@@ -221,7 +232,7 @@ void Output_CalculateObjectLighting(
 
 void Output_LightRoom(ROOM *const room)
 {
-    if (TR_VERSION == 2 && room->light_mode != RLM_NORMAL) {
+    if (room->light_mode != RLM_NORMAL) {
         Output_LightRoomVertices(room);
     } else if (room->flags & RF_DYNAMIC_LIT) {
         for (int32_t i = 0; i < room->mesh.num_vertices; i++) {
@@ -278,9 +289,15 @@ void Output_LightRoom(ROOM *const room)
 
 void Output_InitLight(void)
 {
-    // TODO: consolidate into Output_Init once common.
     if (m_DynamicLights == nullptr) {
         m_DynamicLights = Vector_Create(sizeof(LIGHT));
+    }
+
+    for (int32_t i = 0; i < M_LIGHT_CYCLE; i++) {
+        for (int32_t j = 0; j < M_LIGHT_CYCLE; j++) {
+            m_RoomLightTables[i].table[j] = (j - (M_LIGHT_CYCLE / 2)) * i
+                * M_MAX_ROOM_LIGHT_UNIT / (M_LIGHT_CYCLE - 1);
+        }
     }
 }
 
@@ -306,4 +323,61 @@ void Output_AddDynamicLight(
         .falloff.value_1 = falloff,
     };
     Vector_Add(m_DynamicLights, &light);
+}
+
+int32_t Output_GetRoomLightShade(const ROOM_LIGHT_MODE mode)
+{
+    return m_RoomLightShades[mode];
+}
+
+void Output_LightRoomVertices(const ROOM *const room)
+{
+    const M_ROOM_LIGHT_TABLE *const light_table =
+        &m_RoomLightTables[m_RoomLightShades[room->light_mode]];
+    for (int32_t i = 0; i < room->mesh.num_vertices; i++) {
+        ROOM_VERTEX *const vtx = &room->mesh.vertices[i];
+        const int32_t wibble =
+            light_table->table[vtx->light_table_value % M_LIGHT_CYCLE];
+        vtx->light_adder = vtx->light_base + wibble;
+    }
+}
+
+int32_t Output_GetSunsetDuration(void)
+{
+    return 20 * 60 * LOGIC_FPS; // = 20 minutes / 36000 frames
+}
+
+int32_t Output_GetSunsetTimer(void)
+{
+    return m_SunsetTimer;
+}
+
+void Output_SetSunsetTimer(const int32_t timer)
+{
+    m_SunsetTimer = timer;
+}
+
+void Output_SetSunsetEnabled(const bool enabled)
+{
+    m_IsSunsetEnabled = enabled;
+}
+
+void Output_AnimateLights(const int32_t num_frames)
+{
+    m_WibbleOffset += num_frames;
+    m_WibbleOffset %= M_LIGHT_CYCLE;
+    if (TR_VERSION == 2) {
+        m_RoomLightShades[RLM_FLICKER] = Random_GetDraw() % M_LIGHT_CYCLE;
+        m_RoomLightShades[RLM_GLOW] = (M_LIGHT_CYCLE - 1)
+                * (Math_Sin((m_WibbleOffset * DEG_360) / M_LIGHT_CYCLE)
+                   + 0x4000)
+            >> 15;
+
+        if (m_IsSunsetEnabled) {
+            m_SunsetTimer += num_frames;
+            CLAMPG(m_SunsetTimer, Output_GetSunsetDuration());
+            m_RoomLightShades[RLM_SUNSET] = m_SunsetTimer * (M_LIGHT_CYCLE - 1)
+                / Output_GetSunsetDuration();
+        }
+    }
 }
