+++ src/libtrx/include/libtrx/game/output/lights.h
@@ -12,5 +12,12 @@ void Output_CalculateStaticMeshLight(XYZ_32 pos, SHADE shade, const ROOM *room);
 void Output_CalculateObjectLighting(const ITEM *item, const BOUNDS_16 *bounds);
 void Output_LightRoom(ROOM *room);
 
+int32_t Output_GetSunsetDuration(void);
+int32_t Output_GetSunsetTimer(void);
+void Output_SetSunsetTimer(int32_t timer);
+void Output_SetSunsetEnabled(bool enabled);
+
 void Output_ResetDynamicLights(void);
 void Output_AddDynamicLight(XYZ_32 pos, int32_t intensity, int32_t falloff);
+
+void Output_AnimateLights(int32_t num_frames);
