--- src/tr1/game/output/func.c
@@ -1,17 +1,0 @@
-#include "game/viewport.h"
-#include "global/vars.h"
-
-#include <libtrx/debug.h>
-
-int32_t Output_GetRoomLightShade(const ROOM_LIGHT_MODE mode)
-{
-    // TODO: remove
-    ASSERT_FAIL();
-    return 0;
-}
-
-void Output_LightRoomVertices(const ROOM *const room)
-{
-    // TODO: remove
-    ASSERT_FAIL();
-}
